fix: cast BigInt to int in SQL queries to prevent JSON serialization errors
Fixed JSON.stringify errors that were causing 500 errors in dashboard APIs by
adding explicit ::int casts to SUM() operations that return PostgreSQL bigint:

- API /api/history: Cast demandScore and sampleCount in daily history query
- Comparison hub: Cast demandScore and sampleCount in recent history query

These changes ensure numeric values are properly serialized to JSON without
triggering "cannot serialize BigInt" errors. All other APIs already had proper
conversions in place via toSerializableNumber() or Number() functions.

Verified and tested:
- Build compiles successfully
- All API endpoints reviewed for similar issues
- Conversion functions in analytics queries confirmed working

+++ src/app/api/history/index.ts
@@ -54,14 +54,14 @@ async function buildHistoryPayload() {
       )
       SELECT
         day,
-        SUM("bikesRange" + "anchorsRange") AS "demandScore",
+        SUM("bikesRange" + "anchorsRange")::int AS "demandScore",
         AVG("stationAvgOccupancy") AS "avgOccupancy",
         AVG(CASE
           WHEN "stationAvgOccupancy" IS NULL THEN 0.5
           WHEN ABS("stationAvgOccupancy" - 0.5) >= 0.5 THEN 0
           ELSE 1 - (2 * ABS("stationAvgOccupancy" - 0.5))
         END) AS "balanceIndex",
-        SUM("stationSamples") AS "sampleCount"
+        SUM("stationSamples")::int AS "sampleCount"
       FROM daily_station
       GROUP BY day
       ORDER BY day ASC;
